Added search bar and adding stocks. need to work on GUI for this search bar

+++ src/main/java/org/openjfx/hellofx/GUI/EditStockBox.java
@@ -24,12 +24,12 @@ public class EditStockBox extends GridPane implements EventHandler<ActionEvent>{
 	
 	private Stock stock;
 	private Label slabel,sname;
-	private Button delete;
+	private Button delete,add;
 	private int index;
 	private StockEditFrame parent;
 	
 	
-	public EditStockBox(Stock stock,int index,StockEditFrame parent) {
+	public EditStockBox(Stock stock,int index,StockEditFrame parent,int addordelete) {
 		super();
 		this.stock=stock;
 		this.setIndex(index);
@@ -44,16 +44,26 @@ public class EditStockBox extends GridPane implements EventHandler<ActionEvent>{
 		sname = new Label(stock.getName());
 		sname.setId("StockBox-name");
 		//sname.setBorder(new Border(new BorderStroke(Color.RED, BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderWidths.DEFAULT)));
-		delete = new Button("X");
+		if(addordelete==1) {
+			delete = new Button("X");
+			delete.setId("Delete-button");
+			delete.setOnAction(this);
+			GridPane.setConstraints(delete,0,0);
+			GridPane.setRowSpan(delete, 2);
+			GridPane.setHalignment(delete, HPos.CENTER);
+		}
+		else {
+			add = new Button("+");
+			add.setId("Add-button");
+			add.setOnAction(this);
+			GridPane.setConstraints(add,0,0);
+			GridPane.setRowSpan(add, 2);
+			GridPane.setHalignment(add, HPos.CENTER);
+		}
 		//delete.setBorder(new Border(new BorderStroke(Color.RED, BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderWidths.DEFAULT)));
-		delete.setId("Delete-button");
-		delete.setOnAction(this);
 		//delete.setPadding(new Insets(0,0,0,10));
-		GridPane.setConstraints(delete,0,0);
-		GridPane.setRowSpan(delete, 2);
 		GridPane.setConstraints(slabel,1,0);
 		GridPane.setConstraints(sname,1,1);
-		GridPane.setHalignment(delete, HPos.CENTER);
 		GridPane.setHalignment(sname, HPos.CENTER);
 		GridPane.setHalignment(slabel, HPos.CENTER);
 		ColumnConstraints column1 = new ColumnConstraints();
@@ -61,7 +71,12 @@ public class EditStockBox extends GridPane implements EventHandler<ActionEvent>{
 		//column1.setPercentWidth(70-slabel.getText().length()*3);
 		//this.getColumnConstraints().add(column1);
 		this.getColumnConstraints().addAll( new ColumnConstraints( 25 ), column1);
-		this.getChildren().addAll(delete,slabel,sname);
+		if(addordelete==1) {
+			this.getChildren().addAll(delete,slabel,sname);
+		}
+		else {
+			this.getChildren().addAll(add,slabel,sname);
+		}
 		//this.setBorder(new Border(new BorderStroke(Color.YELLOW, BorderStrokeStyle.SOLID, CornerRadii.EMPTY, BorderWidths.DEFAULT)));
 		//this.setGridLinesVisible(true);
 
@@ -86,7 +101,13 @@ public class EditStockBox extends GridPane implements EventHandler<ActionEvent>{
 
 	@Override
 	public void handle(ActionEvent event) {
-		parent.deleteStock(index);
+		if(event.getSource()==delete) {
+			parent.deleteStock(index);
+		}
+		else {
+			parent.addStock(this.stock);
+			parent.refresh();
+		}
 	}
 	
 }
